+++ src/lib/components/jobs/JobLogsModal.svelte
@@ -1,8 +1,10 @@
 <script>
 	import { displayStandardErrorAlert } from '$lib/common/errors';
+	import { extractJobErrorParts } from '$lib/common/job_utilities';
 	import Modal from '../common/Modal.svelte';
 
-	let logs = '';
+	/** @type {Array<{text: string, highlight: boolean}>} */
+	let logParts = [];
 	let errorAlert = undefined;
 	/** @type {Modal} */
 	let modal;
@@ -13,18 +15,19 @@
 	 * @param log {string|null=}
 	 */
 	export async function show(prjId, jobId, log) {
-
 		// remove previous error
 		if (errorAlert) {
 			errorAlert.hide();
 		}
 
+		let logs;
 		if (log) {
 			logs = log;
 		} else {
 			const job = await fetchJob(prjId, jobId);
 			logs = job?.log || '';
 		}
+		logParts = extractJobErrorParts(logs);
 
 		modal.show();
 	}
@@ -60,6 +63,22 @@
 	</svelte:fragment>
 	<svelte:fragment slot="body">
 		<div id="workflowJobLogsError" />
-		<pre>{logs}</pre>
+		<div class="row">
+			<!-- IMPORTANT: do not reindent the pre block, as it will affect the aspect of the log message -->
+			<pre class="ps-0 pe-0" id="workflow-job-logs">
+<!-- -->{#each logParts as part}<div class:highlight={part.highlight} class="ps-3 pe-3">{part.text}
+<!-- --></div>{/each}</pre>
+		</div>
 	</svelte:fragment>
 </Modal>
+
+<style>
+	#workflow-job-logs {
+		/** avoid issues with overflow of inner divs */
+		display: table;
+	}
+	.highlight {
+		font-weight: bold;
+		background-color: #ffe5e5;
+	}
+</style>